refactor(daemon): group auth modules into auth/
