+++ apps/daemon/src/auth/api-token-auth.ts
