+++ apps/daemon/src/auth/desktop-auth.ts
