+++ apps/daemon/src/auth/origin-validation.ts
