+++ apps/daemon/src/auth/tool-tokens.ts
