+++ apps/daemon/tests/api-token-guard.test.ts
@@ -14,7 +14,7 @@
 
 import type http from 'node:http';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { isApiAuthDisabled, isApiTokenMiddlewareEnabled } from '../src/api-token-auth.js';
+import { isApiAuthDisabled, isApiTokenMiddlewareEnabled } from '../src/auth/index.js';
 import { startServer } from '../src/server.js';
 
 const PREVIOUS_TOKEN = process.env.OD_API_TOKEN;
