+++ apps/daemon/tests/origin-validation.test.ts
@@ -10,7 +10,7 @@ import {
   isAllowedBrowserOrigin,
   isLocalSameOrigin,
   isZeroConfigClipperLibraryRequest,
-} from '../src/origin-validation.js';
+} from '../src/auth/index.js';
 
 type TestRequestOptions = {
   origin?: string;
